updated gemm baselines

+++ PIMbench/gemm/baselines/GPU/gemm.cu
@@ -7,19 +7,18 @@
 #include <stdio.h>
 #include <unistd.h>
 #include <stdlib.h>
-#include <math.h>
 #include <iostream>
 #include <cublas_v2.h>
 
-#include "../../../utilBaselines.h"
+#include "utilBaselines.h"
 
 using namespace std;
 
 #define TOLERANCE	200.0f
 
-vector<int32_t> A;
-vector<int32_t> B;
-vector<int32_t> C;
+vector<float> A;
+vector<float> B;
+vector<float> C;
 
 // Params ---------------------------------------------------------------------
 typedef struct Params
